Update dependencies + clean up
