+++ .github/workflows/website-integrity.yml
@@ -23,14 +23,8 @@ jobs:
       - name: Set up env
         uses: the-guild-org/shared-config/setup@v1
 
-      - name: Fetch remote docs
-        run: cd website && pnpm fetch-remote-docs
-
-      - name: Fetch API reference
-        run: cd website && pnpm fetch-api-reference
-
-      - name: Fix pages structure
-        run: cd website && pnpm fix-pages-structure
+      - name: Run pre-build scripts
+        run: pnpm prebuild:ci
 
       - name: Build
         run: pnpm build
