+++ package.json
@@ -2,10 +2,11 @@
   "name": "graph-docs",
   "private": true,
   "version": "1.0.0",
-  "packageManager": "pnpm@10.7.0",
+  "packageManager": "pnpm@10.8.0",
   "scripts": {
     "dev": "turbo run dev",
     "build": "NODE_OPTIONS='--max_old_space_size=8192' turbo run build",
+    "prebuild:ci": "turbo run prebuild:ci",
     "check": "pnpm typecheck && pnpm lint && pnpm prettier:check",
     "check:fix": "pnpm lint:fix; pnpm prettier",
     "lint": "eslint . --ignore-path .gitignore --max-warnings 0",
@@ -20,9 +21,9 @@
   },
   "devDependencies": {
     "@edgeandnode/eslint-config": "^2.0.3",
-    "@types/node": "^22.13.17",
+    "@types/node": "^22.14.0",
     "eslint": "^8.57.1",
-    "eslint-plugin-mdx": "^3.3.1",
+    "eslint-plugin-mdx": "^3.4.0",
     "prettier": "^3.5.3",
     "prettier-plugin-tailwindcss": "^0.6.11",
     "remark-frontmatter": "^5.0.0",
@@ -30,8 +31,8 @@
     "remark-lint-heading-increment": "^4.0.1",
     "remark-lint-no-heading-punctuation": "^4.0.1",
     "remark-lint-restrict-elements": "workspace:*",
-    "turbo": "^2.4.4",
-    "typescript": "^5.8.2"
+    "turbo": "^2.5.0",
+    "typescript": "^5.8.3"
   },
   "resolutions": {
     "@babel/core": "7.26.7"
