+++ packages/og-image/package.json
@@ -16,13 +16,13 @@
     "yoga-wasm-web": "^0.3.3"
   },
   "devDependencies": {
-    "@cloudflare/workers-types": "^4.20250327.0",
+    "@cloudflare/workers-types": "^4.20250407.0",
     "@types/react": "^18.3.20",
     "jest-image-snapshot": "^6.4.0",
     "tsx": "^4.19.3",
-    "typescript": "^5.8.2",
+    "typescript": "^5.8.3",
     "vitest": "^2.1.9",
-    "wrangler": "^3.114.3"
+    "wrangler": "^3.114.4"
   },
   "sideEffects": false
 }
