+++ turbo.json
@@ -10,6 +10,9 @@
       "outputs": ["dist/**"],
       "dependsOn": ["^build"]
     },
+    "prebuild:ci": {
+      "cache": false
+    },
     "typecheck": {
       "dependsOn": ["^build"]
     },
