+++ website/package.json
@@ -8,6 +8,7 @@
     "predev": "tsup ./src/i18n.ts --format esm",
     "build": "rm -rf .next && rm -rf out && next build",
     "prebuild": "pnpm predev",
+    "prebuild:ci": "pnpm fetch-remote-docs && pnpm fetch-api-reference && pnpm fix-pages-structure",
     "postbuild": "next-sitemap --config next-sitemap.config.mjs && node scripts/sitemap-ci.js",
     "typecheck": "tsc",
     "fetch-remote-docs": "tsx scripts/fetch-remote-docs.ts",
@@ -19,13 +20,13 @@
     "@docsearch/react": "^3.9.0",
     "@edgeandnode/common": "^7.0.4",
     "@edgeandnode/gds": "^6.5.12",
-    "@edgeandnode/go": "^9.4.13",
+    "@edgeandnode/go": "^9.4.14",
     "@emotion/react": "^11.14.0",
     "@graphprotocol/contracts": "6.2.1",
     "@pinax/graph-networks-registry": "^0.6.7",
     "@react-hookz/web": "^25.1.0",
     "@readme/httpsnippet": "^11.0.0",
-    "@readme/openapi-parser": "^3.0.1",
+    "@readme/openapi-parser": "^3.0.2",
     "fetch-har": "^11.1.1",
     "lodash": "^4.17.21",
     "mixpanel-browser": "^2.63.0",
@@ -54,7 +55,7 @@
     "@types/react": "^18.3.20",
     "@types/react-dom": "^18.3.6",
     "autoprefixer": "^10.4.21",
-    "fast-xml-parser": "^5.1.0",
+    "fast-xml-parser": "^5.2.0",
     "postcss": "^8.5.3",
     "tailwindcss": "^3.4.17",
     "tsup": "^8.4.0",
