+++ website/src/pages/[locale]/[...404].mdx
@@ -1,5 +1,4 @@
 import { translate } from '@edgeandnode/gds'
-import { RemoteContent } from 'nextra/components'
 import { buildDynamicMDX } from 'nextra/remote'
 
 import { supportedLocales, translations } from '@/i18n'
@@ -22,8 +21,7 @@ hideTableOfContents: true
 hideContentHeader: true
 hideContentFooter: true
 unwrapContent: true
----
-<PageNotFound />`
+---`
   const mdx = await buildDynamicMDX(rawMdx, { codeHighlight: false })
   const { pageMap } = await import(`.next/static/chunks/nextra-page-map-${locale}.mjs`)
   return {
@@ -34,4 +32,4 @@ unwrapContent: true
   }
 }
 
-<RemoteContent components={{ PageNotFound }} />
+<PageNotFound />
