+++ website/src/pages/[locale]/index.mdx
@@ -1,7 +1,6 @@
 {/* We use `index.mdx` instead of `index.tsx` to render the page in the Nextra theme's `Layout` */}
 
 import { translate } from '@edgeandnode/gds'
-import { RemoteContent } from 'nextra/components'
 import { buildDynamicMDX } from 'nextra/remote'
 
 import HomePage from '@/HomePage'
@@ -22,19 +21,19 @@ title: ${t('index.title')}
 hideTableOfContents: true
 hideContentHeader: true
 unwrapContent: true
----
-<HomePage />`
+---`
   const mdx = await buildDynamicMDX(rawMdx, { codeHighlight: false })
   const { pageMap } = await import(`.next/static/chunks/nextra-page-map-${locale}.mjs`)
+  const supportedNetworks = await getSupportedNetworks()
   return {
     props: {
       ...mdx,
       __nextra_pageMap: pageMap,
-      ssg: {
-        supportedNetworks: await getSupportedNetworks(),
-      },
+      supportedNetworks,
     },
   }
 }
 
-<RemoteContent components={{ HomePage }} />
+export default function Page({ supportedNetworks }) {
+  return <HomePage supportedNetworks={supportedNetworks} />
+}
