+++ website/src/pages/[locale]/supported-networks/[id].mdx
@@ -1,10 +1,8 @@
 import { translate } from '@edgeandnode/gds'
-import { RemoteContent } from 'nextra/components'
 import { buildDynamicMDX } from 'nextra/remote'
 
-import { NetworkPage } from '@/supportedNetworks/components'
 import { supportedLocales, translations } from '@/i18n'
-import { getSupportedNetworks } from '@/supportedNetworks'
+import { getSupportedNetworks, NetworkDetailsPage } from '@/supportedNetworks'
 
 export const getStaticPaths = async () => {
   const networks = await getSupportedNetworks()
@@ -34,15 +32,13 @@ export const getStaticProps = async ({ params }) => {
     }
   }
   const rawMdx = `---
-title: ${network.id}
+title: ${network.fullName}
 hideTableOfContents: true
 hideContentHeader: true
 unwrapContent: true
 hideContentFooter: true
 parentRouteKey: supported-networks
----
-# Network Details
-`
+---`
   const mdx = await buildDynamicMDX(rawMdx, { codeHighlight: false })
   const { pageMap } = await import(`.next/static/chunks/nextra-page-map-${locale}.mjs`)
   return {
@@ -50,13 +46,10 @@ parentRouteKey: supported-networks
       ...mdx,
       __nextra_pageMap: pageMap,
       network,
-      ssg: {
-        network,
-      },
     },
   }
 }
 
-export default function NetworkDetailPage({ network }) {
-  return <NetworkPage network={network} />
+export default function Page({ network }) {
+  return <NetworkDetailsPage network={network} />
 }
