+++ website/src/pages/de/supported-networks.mdx
@@ -4,7 +4,7 @@ hideTableOfContents: true
 hideContentHeader: true
 ---
 
-import { getSupportedNetworksStaticProps, SupportedNetworksTable } from '@/supportedNetworks'
+import { getSupportedNetworksStaticProps, NetworksTable } from '@/supportedNetworks'
 import { Heading } from '@/components'
 import { useI18n } from '@/i18n'
 
@@ -14,7 +14,7 @@ export const getStaticProps = getSupportedNetworksStaticProps
   {useI18n().t('index.supportedNetworks.title')}
 </Heading.H2>
 
-<SupportedNetworksTable networks={props.networks} />
+<NetworksTable networks={props.networks} />
 
 - Subgraph Studio verlässt sich auf die Stabilität und Zuverlässigkeit der zugrundeliegenden Technologien, z.B. JSON-RPC, Firehose und Substreams Endpunkte.
 - Subgraphs, die die Gnosis-Kette indizieren, können jetzt mit dem gnosis- Netzwerkidentifikator eingesetzt werden.
