+++ website/src/pages/mr/supported-networks.mdx
@@ -4,7 +4,7 @@ hideTableOfContents: true
 hideContentHeader: true
 ---
 
-import { getSupportedNetworksStaticProps, SupportedNetworksTable } from '@/supportedNetworks'
+import { getSupportedNetworksStaticProps, NetworksTable } from '@/supportedNetworks'
 import { Heading } from '@/components'
 import { useI18n } from '@/i18n'
 
@@ -14,7 +14,7 @@ export const getStaticProps = getSupportedNetworksStaticProps
   {useI18n().t('index.supportedNetworks.title')}
 </Heading.H2>
 
-<SupportedNetworksTable networks={props.networks} />
+<NetworksTable networks={props.networks} />
 
 - Subgraph Studio relies on the stability and reliability of the underlying technologies, for example JSON-RPC, Firehose and Substreams endpoints.
 - Subgraphs indexing Gnosis Chain can now be deployed with the `gnosis` network identifier.
