--- website/src/supportedNetworks/components/index.ts
@@ -1,7 +1,0 @@
-export * from './EmptySearchResults'
-export * from './NetworkDetailsPage'
-export * from './NetworkFilters'
-export * from './NetworkPage'
-export * from './NetworkRow'
-export * from './NetworksTable'
-export * from './SkeletonRow'
