+++ website/src/supportedNetworks/utils.ts
@@ -1,44 +1,35 @@
-import { NetworkType } from '@pinax/graph-networks-registry'
+import { NetworksRegistry, NetworkType } from '@pinax/graph-networks-registry'
 
-// Networks that should use the "mono" icon variant
+// Networks that should use the "mono" icon variant (TODO: add this feature to web3icons?)
 export const MONO_ICON_NETWORKS = [
+  'arweave-mainnet',
+  'autonomys-taurus',
+  'expchain-testnet',
+  'fraxtal',
+  'lens',
+  'lens-testnet',
+  'linea',
+  'linea-sepolia',
+  'lumia',
+  'mbase',
+  'megaeth-testnet',
+  'soneium',
+  'soneium-testnet',
+  'sonic',
+  'stellar',
   'vana',
   'vana-moksha',
   'xlayer-mainnet',
   'xlayer-sepolia',
   'zksync-era',
   'zksync-era-sepolia',
-  'sonic',
-  'soneium-testnet',
-  'soneium',
-  'linea-sepolia',
-  'linea',
-  'lens-testnet',
-  'expchain-testnet',
-  'autonomys-taurus',
-  'arweave-mainnet',
-  'fraxtal',
-  'lumia',
-  'mbase',
-  'megaeth-testnet',
 ]
-
-// TODO(@hayderkg, @benface, @0xa3k5): Add network icons to GDS for these networks
-export const MISSING_ICON_NETWORKS: string[] = []
 
 // Networks with Token API support (TODO: remove once the registry has this information)
 export const TOKEN_API_NETWORKS = ['mainnet', 'base', 'bsc', 'arbitrum-one', 'matic', 'optimism']
 
 export const getIconVariant = (networkId: string): 'mono' | 'branded' => {
   return MONO_ICON_NETWORKS.includes(networkId) ? 'mono' : 'branded'
-}
-
-export const shouldShowSkeleton = (networkId: string): boolean => {
-  return MISSING_ICON_NETWORKS.includes(networkId) || !networkId
-}
-
-export const supportsTokenAPI = (networkId: string): boolean => {
-  return TOKEN_API_NETWORKS.includes(networkId)
 }
 
 export interface Network {
@@ -109,4 +100,17 @@ export const processNetworksData = (networks: NetworkData[]): ProcessedNetwork[]
 
 export const isEVMNetwork = (network: Network | NetworkData): boolean => {
   return network.caip2Id.startsWith('eip155:')
+}
+
+export async function getSupportedNetworks(): Promise<ProcessedNetwork[]> {
+  const registry = await NetworksRegistry.fromLatestVersion()
+  return processNetworksData(registry.networks)
+}
+
+export async function getSupportedNetworksStaticProps() {
+  return {
+    props: {
+      networks: await getSupportedNetworks(),
+    },
+  }
 }
